Move authprovider to pkg/common/oras
