+++ pkg/common/oras/authprovider/authprovider.go
