+++ pkg/common/oras/authprovider/authprovider_test.go
