+++ pkg/common/oras/authprovider/authproviderconfig.go
