+++ pkg/common/oras/authprovider/authproviderfactory.go
