+++ pkg/common/oras/authprovider/authproviderfactory_test.go
