+++ pkg/common/oras/authprovider/aws/ecrauthtypes.go
