+++ pkg/common/oras/authprovider/azure/const.go
