+++ pkg/referrerstore/oras/oras.go
@@ -37,14 +37,14 @@ import (
 
 	ratifyconfig "github.com/deislabs/ratify/config"
 	"github.com/deislabs/ratify/pkg/common"
+	"github.com/deislabs/ratify/pkg/common/oras/authprovider"
+	_ "github.com/deislabs/ratify/pkg/common/oras/authprovider/aws"
+	_ "github.com/deislabs/ratify/pkg/common/oras/authprovider/azure"
 	"github.com/deislabs/ratify/pkg/homedir"
 	"github.com/deislabs/ratify/pkg/ocispecs"
 	"github.com/deislabs/ratify/pkg/referrerstore"
 	"github.com/deislabs/ratify/pkg/referrerstore/config"
 	"github.com/deislabs/ratify/pkg/referrerstore/factory"
-	"github.com/deislabs/ratify/pkg/referrerstore/oras/authprovider"
-	_ "github.com/deislabs/ratify/pkg/referrerstore/oras/authprovider/aws"
-	_ "github.com/deislabs/ratify/pkg/referrerstore/oras/authprovider/azure"
 	"github.com/opencontainers/go-digest"
 	"github.com/sirupsen/logrus"
 )
